Add layout system (#861)
* Refactor App to use layouts

* Lint fix

* Restructure pool pages

* Fix BlankLayout

* Remove redundant class

* Clean up DefaultLayout

* Adjust page component naming
In Nuxt they have this pattern where the pages directory matches the available routes, where the file names describe the page paths.

* Re-add copper module

* FIx import reference

* Add layouts readme

* Rename case sensitive file
