+++ src/components/layouts/BlankLayout.vue
@@ -1,0 +1,5 @@
+<script setup lang="ts"></script>
+
+<template>
+  <router-view :key="$route.path" />
+</template>
