+++ src/components/layouts/index.ts
@@ -1,0 +1,2 @@
+export { default as DefaultLayout } from './DefaultLayout.vue';
+export { default as BlankLayout } from './BlankLayout.vue';
