+++ src/pages/pool/_id.vue
