+++ src/plugins/router.ts
@@ -1,25 +1,41 @@
 import { createRouter, createWebHashHistory, RouteRecordRaw } from 'vue-router';
-import Home from '@/pages/Home.vue';
-import Pool from '@/pages/Pool.vue';
-import LiquidityMining from '@/pages/LiquidityMining.vue';
-import Trade from '@/pages/Trade.vue';
+import HomePage from '@/pages/index.vue';
+import PoolPage from '@/pages/pool/_id.vue';
+import LiquidityMiningPage from '@/pages/liquidity-mining.vue';
+import TradePage from '@/pages/trade.vue';
 
 const routes: RouteRecordRaw[] = [
-  { path: '/', name: 'home', component: Home },
-  { path: '/trade/:assetIn?/:assetOut?', name: 'trade', component: Trade },
+  {
+    path: '/',
+    name: 'home',
+    component: HomePage
+  },
+  {
+    path: '/trade/:assetIn?/:assetOut?',
+    name: 'trade',
+    component: TradePage
+  },
   {
     path: '/swap/:assetIn?/:assetOut?',
     redirect: to => {
       return `/trade${to.path.split('/swap')[1]}`;
     }
   },
-  { path: '/pool/:id', name: 'pool', component: Pool },
+  {
+    path: '/pool/:id',
+    name: 'pool',
+    component: PoolPage
+  },
   {
     path: '/liquidity-mining',
     name: 'liquidity-mining',
-    component: LiquidityMining
+    component: LiquidityMiningPage
   },
-  { path: '/:pathMatch(.*)*', name: 'not-found', redirect: '/' }
+  {
+    path: '/:pathMatch(.*)*',
+    name: 'not-found',
+    redirect: '/'
+  }
 ];
 
 const router = createRouter({
